flask_restful_02_basic_auth/requirements.txt
flask
flask-sqlalchemy
flask-restful
flask-httpauth
